% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\alias{monitor.data.frame}
\title{Title}
\usage{
monitor(x, ...)

\method{monitor}{data.frame}(x, watch_vars = NULL, group_var = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
