% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{print.track_distribution}
\alias{print.track_distribution}
\title{Title}
\usage{
\method{print}{track_distribution}(x, width = max(20, options()$width - 35),
  ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
