% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{track_distribution}
\alias{track_distribution}
\title{Title}
\usage{
track_distribution(m, ..., distance_method = "jensen-shannon",
  method = "fix", options = list(), focus = m$watch_vars, role = NULL,
  initialized = FALSE, skip = FALSE, id = rand_id("track_distribution"))
}
\arguments{
\item{id}{}
}
\description{
Title
}
